Implement several new commands, and refactor code
The new commands go hand in hand with the refactoring: I changed the getBadMemes function into a generic getRedditPictures function and instead seperated the different commands into their own json parsing functions to get a subreddit url to pass to the getRedditPicctures function.
With that, I added !cat, !dog, and !aww, which gets cat pictures, dog pictures and animal pictures in general.